+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.9.3] - 2024-10-16
+
 ### Changed
 
 - [fj-doc-maven-plugin] support for asciidoc doc handler (fj-doc-version 8.8.7+)